domain_bot/requirements.txt
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.36.0
llama-cpp-python==0.2.82
transformers==4.42.4
accelerate==0.33.0
bitsandbytes==0.43.1
peft==0.11.1
torch==2.3.1
scikit-learn==1.5.1
nltk==3.8.1
tqdm==4.66.4
